+++ src/HorseCollisionMod/Log.lua
@@ -12,7 +12,7 @@
 --
 -- @module HorseCollisionMod.Log
 -- @author jrandall54
--- @release 4.3.7
+-- @release 4.3.8
 
 --- The engine clock in milliseconds.
 --